+++ src/test/java/cart/domain/DiscountPriceCalculatorTest.java
@@ -1,6 +1,6 @@
 package cart.domain;
 
-import cart.domain.discount_strategy.DiscountPriceCalculator;
+import cart.domain.discount.strategy.DiscountPriceCalculator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
